Move handling of before/after terms from post-parameters callback to default view
Handling the before/after terms on post queries was the only thing
view-parameters callbacks were being used for, so this means
view-parameters callbacks are now unused and can also be deleted.
