+++ packages/lesswrong/lib/collections/posts/index.ts
@@ -1,5 +1,4 @@
 export * from './collection';
-import './parameters';
 import './views';
 import './helpers';
 import './permissions';
